Move PrettifyCustomEmoji app/engine to app/utils
