+++ app/utils/formatting.go
@@ -2,11 +2,27 @@ package utils
 
 import (
 	"fmt"
+	"regexp"
 	"strings"
 	"time"
 
 	"github.com/fiffu/arisa3/lib"
 )
+
+// Misc values
+var (
+	CustomEmojiRegex *regexp.Regexp = regexp.MustCompile(`(<a?)?:\w+:(\d{18}>)?`)
+)
+
+// PrettifyCustomEmoji converts "<:birb:924875584004321361>" into ":birb:"
+func PrettifyCustomEmoji(str string) string {
+	replacer := func(s string) string {
+		start := strings.Index(s, ":")
+		end := strings.LastIndex(s, ":")
+		return s[start : end+1]
+	}
+	return CustomEmojiRegex.ReplaceAllStringFunc(str, replacer)
+}
 
 func EscapeMarkdown(text string) string {
 	markdownSymbols := "\\*>_:~`|"
